% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{calculateWeights}
\alias{calculateWeights}
\title{calculateWeights}
\usage{
calculateWeights(criteriaNames, bestToOthers, othersToWorst)
}
\arguments{
\item{criteriaNames}{Names of the criteria}

\item{bestToOthers}{Vector of pairwise comparisons. Best criterion should be 1, others <2, 9>.}

\item{othersToWorst}{Vector of pairwise comparisons. Worst criterion should be 1, others <2, 9>.}
}
\value{
Result that consist of \code{criteriaNames}, \code{criteriaWeights}, \code{consistencyRatio} and a model that was used to calculate weights.
}
\description{
Implementation based on https://doi.org/10.1016/j.omega.2015.12.001.
Calculates weights of the criteria using a linear model.
Steps:
1. Build model (consists of validating model and constructing necessary constraints for LP problem).
2. Solve LP problem.
3. Calculate consistency ratio.
}
\examples{
criteriaNames <- c("quality", "price", "comfort", "safety", "style")
bestToOthers <- c(2, 1, 4, 2, 8)
othersToWorst <- c(4, 8, 2, 4, 1)
calculateWeights(criteriaNames, bestToOthers, othersToWorst)
}
